+++ Near/ref-contracts/ref-exchange/tests/fuzzy/utils.rs
@@ -18,7 +18,7 @@ use crate::fuzzy::constants::*;
 
 near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
     TEST_TOKEN_WASM_BYTES => "../res/test_token.wasm",
-    EXCHANGE_WASM_BYTES => "../res/ref_exchange_release.wasm",
+    EXCHANGE_WASM_BYTES => "../res/ref_exchange.wasm",
 }
 
 /**
@@ -372,6 +372,7 @@ pub fn setup_stable_pool_with_liquidity_and_operators(
         pool.add_stable_swap_pool(
             (&token_contracts).into_iter().map(|x| x.valid_account_id()).collect(), 
             decimals,
+            pool_fee,
             amp
         ),
         deposit = to_yocto("1"))