+++ Near/scripts/Nodejs/nearConfig.js
@@ -2,7 +2,8 @@ exports.CONTRACTS = {
     exchange: "png-exchange-v1.testnet",
     farming: "png-farm-v1.testnet",
     token: "png-token-v1.testnet",
-    xToken: "png-xtoken-v1.testnet"
+    xToken: "png-xtoken-v1.testnet",
+    vampire: "png-vampire-v2.testnet"
 };
 
 exports.OWNER = "0xsimo.testnet"
@@ -34,6 +35,17 @@ exports.FARMS = [
         session_interval: 1000,
         amountRewardToken: "0"
     }
+]
+
+exports.VFARMS = [
+    {
+        poolId: 0,
+        rewardToken: "png-token-v1.testnet",
+        start_at: 1653512400,
+        reward_per_session: "1",
+        session_interval: 1000,
+        amountRewardToken: "10000000000000000000"
+    },
 ]