+++ Near/ref-contracts/ref-farming/release_notes.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+### Version 1.1.0
+1. Add state control by owner;
+
+### Version 1.0.4
+1. Return promise in withdraw reward and seed;
+
+### Version 1.0.3
+1. Fix bug on modify_seed_min_deposit;
+
 ### Version 1.0.2
 1. 80T gas for seed withdraw resolve, 10T gas for reward withdraw resolve;
 2. add get_user_storage_state interface;